# docs/openapi-v2.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REVENANT Control Plane API",
    "description": "Autonomous Full-Spectrum AI Red Teaming & Security Operations Platform",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Health Check",
        "description": "System health check endpoint.",
        "operationId": "health_check_api_v1_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Api V1 Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create And Start Campaign",
        "description": "Launch an autonomous security campaign.\nEnforces Layer 1 Scope Check on intake before any engine starts.",
        "operationId": "create_and_start_campaign_api_v1_campaigns_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign",
        "description": "Get mission summary and status.",
        "operationId": "get_campaign_api_v1_campaigns__campaign_id__get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/findings": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign Findings",
        "description": "Retrieve all findings discovered during a campaign.",
        "operationId": "get_campaign_findings_api_v1_campaigns__campaign_id__findings_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Finding"
                  },
                  "title": "Response Get Campaign Findings Api V1 Campaigns  Campaign Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/assets": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get Campaign Assets",
        "description": "Retrieve all assets discovered during a campaign.",
        "operationId": "get_campaign_assets_api_v1_campaigns__campaign_id__assets_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HostAsset"
                  },
                  "title": "Response Get Campaign Assets Api V1 Campaigns  Campaign Id  Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/report": {
      "get": {
        "tags": [
          "Reporting"
        ],
        "summary": "Get Campaign Report",
        "description": "Download or view campaign report in Markdown, HTML, SARIF v2.1.0, JSON, Faraday, or Dradis.",
        "operationId": "get_campaign_report_api_v1_campaigns__campaign_id__report_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(markdown|html|sarif|json|faraday|dradis)$",
              "description": "Desired export format",
              "default": "markdown",
              "title": "Format"
            },
            "description": "Desired export format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CampaignCreateRequest": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target",
            "example": "http://localhost:3000"
          },
          "scope": {
            "$ref": "#/components/schemas/ScopeManifest"
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "default": 25
          }
        },
        "type": "object",
        "required": [
          "target",
          "scope"
        ],
        "title": "CampaignCreateRequest"
      },
      "CampaignResponse": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "target",
          "status",
          "created_at",
          "summary"
        ],
        "title": "CampaignResponse"
      },
      "CodeLocation": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "end_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Line"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "commit_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Hash"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "start_line"
        ],
        "title": "CodeLocation",
        "description": "Specific source code or repository location where a vulnerability occurs."
      },
      "EndpointInfo": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": "GET"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "tech_stack": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tech Stack"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Headers"
          },
          "parameters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Parameters"
          },
          "forms": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Forms"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "EndpointInfo",
        "description": "Web / HTTP endpoint discovery information."
      },
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "cve": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cve"
          },
          "cwe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwe"
          },
          "cvss_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cvss Score"
          },
          "mitre_attack_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mitre Attack Ids"
          },
          "owasp_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owasp Category"
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "code_location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CodeLocation"
              },
              {
                "type": "null"
              }
            ]
          },
          "secret_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Type"
          },
          "reproduction_steps": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reproduction Steps"
          },
          "evidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence"
          },
          "request_proof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Proof"
          },
          "response_proof": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Proof"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation"
          },
          "raw_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Data"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          },
          "verified_valid": {
            "type": "boolean",
            "title": "Verified Valid",
            "default": false
          },
          "false_positive": {
            "type": "boolean",
            "title": "False Positive",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "severity",
          "target",
          "tool"
        ],
        "title": "Finding",
        "description": "Standardized vulnerability finding across all 66 security tools."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HostAsset": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          },
          "ports": {
            "items": {
              "$ref": "#/components/schemas/PortInfo"
            },
            "type": "array",
            "title": "Ports"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/EndpointInfo"
            },
            "type": "array",
            "title": "Endpoints"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "cloud_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Provider"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "first_seen": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen"
          }
        },
        "type": "object",
        "title": "HostAsset",
        "description": "Discovered network host or domain asset."
      },
      "PortInfo": {
        "properties": {
          "port": {
            "type": "integer",
            "title": "Port"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "default": "tcp"
          },
          "state": {
            "type": "string",
            "title": "State",
            "default": "open"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "banner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banner"
          }
        },
        "type": "object",
        "required": [
          "port"
        ],
        "title": "PortInfo",
        "description": "Network port scan result."
      },
      "ScopeManifest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "default-scope"
          },
          "allowed_cidrs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Cidrs"
          },
          "allowed_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Domains"
          },
          "allowed_wildcards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Wildcards"
          },
          "allowed_hosts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Hosts"
          },
          "allowed_repos": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Repos"
          },
          "allowed_cloud_accounts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Cloud Accounts"
          },
          "allowed_container_images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Container Images"
          },
          "allowed_identity_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Identity Domains"
          },
          "denied_cidrs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Denied Cidrs"
          },
          "denied_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Denied Domains"
          },
          "allow_cloud_metadata": {
            "type": "boolean",
            "title": "Allow Cloud Metadata",
            "default": false
          },
          "require_explicit_scope": {
            "type": "boolean",
            "title": "Require Explicit Scope",
            "default": true
          },
          "max_parallel_active": {
            "type": "integer",
            "title": "Max Parallel Active",
            "default": 20
          }
        },
        "type": "object",
        "title": "ScopeManifest",
        "description": "Authorization manifest defining allowed and strictly forbidden targets."
      },
      "Severity": {
        "type": "string",
        "enum": [
          "INFO",
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "Severity"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}